refactor(ui): colocate AJOK PDF rule-set renderers
  - move legacy PDF section renderers into their owning rule-period folders
  - update registry, public exports, and tests to use rule-set-local paths
