+++ apps/web/lib/public/beagle/trials/pdf/__tests__/trial-dog-pdf-allekirjoitukset.test.ts
@@ -1,6 +1,6 @@
 import type { PDFFont, PDFPage } from "pdf-lib";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { drawTrialDogPdfAllekirjoitukset } from "../internal/allekirjoitukset";
+import { drawTrialDogPdfAllekirjoitukset } from "../rule-sets/legacy-2011-2023/allekirjoitukset";
 
 describe("drawTrialDogPdfAllekirjoitukset", () => {
   const page = {
