+++ apps/web/lib/public/beagle/trials/pdf/rule-sets/legacy-2005-2011/index.ts
@@ -1,7 +1,7 @@
 // Orchestrates the AJOK 2005-2011 renderer. Sections are enabled one at a time
 // while coordinates are tuned against the template.
-import { drawLegacy2005To2011KokeenTiedot } from "./legacy-2005-2011/kokeen-tiedot";
-import type { TrialDogPdfRenderContext } from "./types";
+import type { TrialDogPdfRenderContext } from "../types";
+import { drawLegacy2005To2011KokeenTiedot } from "./kokeen-tiedot";
 
 export function renderLegacy2005To2011TrialDogPdfFields({
   input,
