+++ apps/web/lib/public/beagle/trials/pdf/rule-sets/legacy-2011-2023/koe-erat-common.ts
